removed legacy parser, renamed file for MT reimplementation
